Fixing download all resources
Fix the download_all_resources with the error on XNAT that randomly unzip a zip file when uploading it. IF only one file when download_all_resources, doesn't zip it, just download it and if it's a zip file, unzip it ELSE if more than one resource, download all of them as a zip and unzip it.